Refactor AbstractBitcoinNetParams: Bifurcate into AbstractBitcoinCashParams and AbstractBitcoinNetParams. Bip47AppKit: Add getPeerGroup method, Save wallet file after removing tx, set default max connections before adding ips. PeerGroup: Improve logging
