+++ core/src/main/java/org/bitcoinj/core/TransactionBroadcast.java
@@ -132,7 +132,7 @@ public class TransactionBroadcast {
             // We will send the tx simultaneously to half the connected peers and wait to hear back from at least half
             // of the other half, i.e., with 4 peers connected we will send the tx to 2 randomly chosen peers, and then
             // wait for it to show up on one of the other two. This will be taken as sign of network acceptance. As can
-            // be seen, 4 peers is probably too little - it doesn't taken many broken peers for tx propagation to have
+            // be seen, 4 peers is probably too little - it doesn't take many broken peers for tx propagation to have
             // a big effect.
             List<Peer> peers = peerGroup.getConnectedPeers();    // snapshots
             // Prepare to send the transaction by adding a listener that'll be called when confidence changes.
@@ -158,7 +158,7 @@ public class TransactionBroadcast {
                 try {
                     peer.sendMessage(tx);
                     // We don't record the peer as having seen the tx in the memory pool because we want to track only
-                    // how many peers announced to us.
+                    // how many peers announcgit ed to us.
                 } catch (Exception e) {
                     log.error("Caught exception sending to {}", peer, e);
                 }
