--- core/src/main/java/org/bitcoinj/params/AbstractBitcoinNetParams.java
@@ -1,178 +1,0 @@
-/*
- * Copyright 2013 Google Inc.
- * Copyright 2015 Andreas Schildbach
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.bitcoinj.params;
-
-import java.math.BigInteger;
-import java.util.concurrent.TimeUnit;
-
-import org.bitcoinj.core.*;
-import org.bitcoinj.utils.MonetaryFormat;
-import org.bitcoinj.store.BlockStore;
-import org.bitcoinj.store.BlockStoreException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import com.google.common.base.Stopwatch;
-
-/**
- * Parameters for Bitcoin-like networks.
- */
-public abstract class AbstractBitcoinNetParams extends NetworkParameters {
-    /**
-     * Scheme part for Bitcoin URIs.
-     */
-    public static final String BITCOIN_SCHEME = "bitcoin";
-
-    private static final Logger log = LoggerFactory.getLogger(AbstractBitcoinNetParams.class);
-
-    public AbstractBitcoinNetParams() {
-        super();
-    }
-
-    /**
-     * Checks if we are at a difficulty transition point.
-     * @param storedPrev The previous stored block
-     * @return If this is a difficulty transition point
-     */
-    protected boolean isDifficultyTransitionPoint(StoredBlock storedPrev) {
-        return ((storedPrev.getHeight() + 1) % this.getInterval()) == 0;
-    }
-
-    @Override
-    public void checkDifficultyTransitions(final StoredBlock storedPrev, final Block nextBlock,
-       final BlockStore blockStore) throws VerificationException, BlockStoreException {
-        Block prev = storedPrev.getHeader();
-
-        // Is this supposed to be a difficulty transition point?
-        if (!isDifficultyTransitionPoint(storedPrev)) {
-
-            // No ... so check the difficulty didn't actually change.
-            if (nextBlock.getDifficultyTarget() != prev.getDifficultyTarget())
-                throw new VerificationException("Unexpected change in difficulty at height " + storedPrev.getHeight() +
-                        ": " + Long.toHexString(nextBlock.getDifficultyTarget()) + " vs " +
-                        Long.toHexString(prev.getDifficultyTarget()));
-            return;
-        }
-
-        // We need to find a block far back in the chain. It's OK that this is expensive because it only occurs every
-        // two weeks after the initial block chain download.
-        final Stopwatch watch = Stopwatch.createStarted();
-        StoredBlock cursor = blockStore.get(prev.getHash());
-        for (int i = 0; i < this.getInterval() - 1; i++) {
-            if (cursor == null) {
-                // This should never happen. If it does, it means we are following an incorrect or busted chain.
-                throw new VerificationException(
-                        "Difficulty transition point but we did not find a way back to the genesis block.");
-            }
-            cursor = blockStore.get(cursor.getHeader().getPrevBlockHash());
-        }
-        watch.stop();
-        if (watch.elapsed(TimeUnit.MILLISECONDS) > 50)
-            log.info("Difficulty transition traversal took {}", watch);
-
-        Block blockIntervalAgo = cursor.getHeader();
-        int timespan = (int) (prev.getTimeSeconds() - blockIntervalAgo.getTimeSeconds());
-        // Limit the adjustment step.
-        final int targetTimespan = this.getTargetTimespan();
-        if (timespan < targetTimespan / 4)
-            timespan = targetTimespan / 4;
-        if (timespan > targetTimespan * 4)
-            timespan = targetTimespan * 4;
-
-        BigInteger newTarget = Utils.decodeCompactBits(prev.getDifficultyTarget());
-        newTarget = newTarget.multiply(BigInteger.valueOf(timespan));
-        newTarget = newTarget.divide(BigInteger.valueOf(targetTimespan));
-
-        if (newTarget.compareTo(this.getMaxTarget()) > 0) {
-            log.info("Difficulty hit proof of work limit: {}", newTarget.toString(16));
-            newTarget = this.getMaxTarget();
-        }
-
-        int accuracyBytes = (int) (nextBlock.getDifficultyTarget() >>> 24) - 3;
-        long receivedTargetCompact = nextBlock.getDifficultyTarget();
-
-        // The calculated difficulty is to a higher precision than received, so reduce here.
-        BigInteger mask = BigInteger.valueOf(0xFFFFFFL).shiftLeft(accuracyBytes * 8);
-        newTarget = newTarget.and(mask);
-        long newTargetCompact = Utils.encodeCompactBits(newTarget);
-
-        if (newTargetCompact != receivedTargetCompact)
-            throw new VerificationException("Network provided difficulty bits do not match what was calculated: " +
-                    Long.toHexString(newTargetCompact) + " vs " + Long.toHexString(receivedTargetCompact));
-    }
-
-    @Override
-    public Coin getMaxMoney() {
-        return MAX_MONEY;
-    }
-
-    @Override
-    public Coin getMinNonDustOutput() {
-        return Transaction.MIN_NONDUST_OUTPUT;
-    }
-
-    @Override
-    public Coin getReferenceDefaultMinTxFee() {
-        return Transaction.REFERENCE_DEFAULT_MIN_TX_FEE;
-    }
-
-    @Override
-    public int getMaxBlockSize() {
-        return Block.MAX_BLOCK_SIZE;
-    }
-
-    @Override
-    public int getMaxBlockSigops() {
-        return Block.MAX_BLOCK_SIGOPS;
-    }
-
-    @Override
-    public boolean getUseForkId() {
-        return false;
-    }
-
-    @Override
-    public Coin getDefaultTxFee() {
-        return Transaction.DEFAULT_TX_FEE;
-    }
-
-    @Override
-    public MonetaryFormat getMonetaryFormat() {
-        return new MonetaryFormat();
-    }
-
-    @Override
-    public int getProtocolVersionNum(final ProtocolVersion version) {
-        return version.getBitcoinProtocolVersion();
-    }
-
-    @Override
-    public BitcoinSerializer getSerializer(boolean parseRetain) {
-        return new BitcoinSerializer(this, parseRetain);
-    }
-
-    @Override
-    public String getUriScheme() {
-        return BITCOIN_SCHEME;
-    }
-
-    @Override
-    public boolean hasMaxMoney() {
-        return true;
-    }
-}
