+++ core/src/main/java/org/bitcoinj/params/UnitTestParams.java
@@ -24,7 +24,7 @@ import java.math.BigInteger;
  * Network parameters used by the bitcoinj unit tests (and potentially your own). This lets you solve a block using
  * {@link org.bitcoinj.core.Block#solve()} by setting difficulty to the easiest possible.
  */
-public class UnitTestParams extends AbstractBitcoinNetParams {
+public class UnitTestParams extends AbstractBitcoinCoreParams {
     public static final int UNITNET_MAJORITY_WINDOW = 8;
     public static final int TESTNET_MAJORITY_REJECT_BLOCK_OUTDATED = 6;
     public static final int TESTNET_MAJORITY_ENFORCE_BLOCK_UPGRADE = 4;
